requirements.txt
aniso8601==9.0.1
astroid==2.6.6
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-mongoengine==1.0.0
flask-restx==0.5.0
Flask-WTF==0.15.1
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mongoengine==0.23.1
pycodestyle==2.7.0
pycparser==2.20
pylint==2.9.6
pymongo==3.12.0
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
